Создает модули mocks.js и util.js
