--- js/ckeck-max-length.js
@@ -1,9 +1,0 @@
-function checkMaxLength (string, maxString) {
-  if (typeof string === 'string' && string.length <= maxString) {
-    return true;
-  } else {
-    return false;
-  }
-}
-
-export { checkMaxLength };
