+++ js/main.js
@@ -1,12 +1,4 @@
-// Функция, возвращающая случайное целое число из переданного диапазона включительно.
-import { randomNumber } from './random-number.js';
-// Функция для проверки максимальной длины строки.
-// import { checkMaxLength } from './ckeck-max-length.js';
-// Функция возвращает случайное описание из массива
-import { getDescription } from './get-description.js';
-// Функция создает обект
-import { createPhoto } from './create-photo.js';
-
-import { photoCards } from './util.js';
+import {photoCards} from './mocks.js';
+import './util.js';
 
 window.console.log(photoCards);
