--- js/random-number.js
@@ -1,10 +1,0 @@
-function randomNumber (minNumber, maxNumber) {
-  if (minNumber < 0 || maxNumber < 0 || minNumber >= maxNumber) {
-    return NaN;
-  }
-
-  const number = minNumber + Math.random() * (maxNumber - minNumber);
-  return Math.round(number);
-}
-
-export {randomNumber};
